+++ core/src/com/daniebker/balloongame/InputCommand.java
@@ -1,6 +1,0 @@
-package com.daniebker.balloongame;
-
-public interface InputCommand {
-
-    void execute(Player player);
-}