+++ core/src/com/daniebker/balloongame/input/InputCommand.java
@@ -1,0 +1,8 @@
+package com.daniebker.balloongame.input;
+
+import com.daniebker.balloongame.Player;
+
+public interface InputCommand {
+
+    void execute(Player player);
+}