+++ core/src/com/daniebker/balloongame/input/InputHandler.java
@@ -1,4 +1,4 @@
-package com.daniebker.balloongame;
+package com.daniebker.balloongame.input;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;