+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2021-03-25
+
+* Implement command pattern for input
+* Don't let player move too far left
+
 ## 2021-03-20
 
 * Initial Player Sprite